Move flow execution analysis to its own class
